Rename CassandraServer to TestingCassandraServer
