+++ plugin/trino-cassandra/src/test/java/io/trino/plugin/cassandra/TestCassandraConnector.java
@@ -99,7 +99,7 @@ public class TestCassandraConnector
     private static final ConnectorSession SESSION = TestingConnectorSession.builder()
             .setPropertyMetadata(new CassandraSessionProperties(new CassandraClientConfig()).getSessionProperties())
             .build();
-    private CassandraServer server;
+    private TestingCassandraServer server;
     protected String database;
     protected SchemaTableName table;
     protected SchemaTableName tableForDelete;
@@ -113,7 +113,7 @@ public class TestCassandraConnector
     public void setup()
             throws Exception
     {
-        this.server = new CassandraServer();
+        this.server = new TestingCassandraServer();
 
         String keyspace = "test_connector";
         createTestTables(server.getSession(), keyspace, DATE);
